# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user-settings": {
      "get": {
        "operationId": "UserSettingsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Prečítať nastavenia používateľa",
        "tags": [
          "User Settings"
        ]
      },
      "patch": {
        "operationId": "UserSettingsController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserSettingsDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upraviť nastavenia používateľa",
        "tags": [
          "User Settings"
        ]
      }
    },
    "/user-settings/heartbeat": {
      "post": {
        "operationId": "UserSettingsController_recordHeartbeat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatTriggerDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Heartbeat spustený"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vyvolať heartbeat ping",
        "tags": [
          "User Settings"
        ]
      }
    },
    "/user-settings/heartbeat-guardians": {
      "get": {
        "operationId": "UserSettingsController_listHeartbeatGuardians",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartbeatGuardianResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all guardians assigned to the Heart-Beat protocol",
        "tags": [
          "User Settings"
        ]
      },
      "post": {
        "operationId": "UserSettingsController_addHeartbeatGuardian",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddHeartbeatGuardianDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Assigned successfully"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a guardian to the Heart-Beat protocol with a specific priority",
        "tags": [
          "User Settings"
        ]
      }
    },
    "/user-settings/heartbeat-guardians/{guardianId}": {
      "delete": {
        "operationId": "UserSettingsController_removeHeartbeatGuardian",
        "parameters": [
          {
            "name": "guardianId",
            "required": true,
            "in": "path",
            "description": "Guardian ID (cuid)",
            "schema": {
              "example": "cku6l6xhc0000q5l5zj9m1xyz",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed successfully or no-op"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a guardian from the Heart-Beat protocol",
        "tags": [
          "User Settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Phoenix API",
    "description": "Dokumentácia API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "object",
            "example": "Validation failed"
          },
          "code": {
            "type": "string",
            "example": "USR_001"
          },
          "path": {
            "type": "string",
            "example": "/user-settings/heartbeat-guardians"
          },
          "timestamp": {
            "type": "string",
            "example": "2025-08-14T22:54:16.250Z"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "UserSettingsDto": {
        "type": "object",
        "properties": {
          "heartbeatIntervalDays": {
            "type": "number",
            "example": 30
          },
          "isHeartbeatActive": {
            "type": "boolean",
            "example": false
          },
          "lastHeartbeatAt": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "notificationChannels": {
            "type": "array",
            "example": [
              "email"
            ],
            "items": {
              "type": "string",
              "enum": [
                "email"
              ]
            }
          }
        },
        "required": [
          "heartbeatIntervalDays",
          "isHeartbeatActive",
          "lastHeartbeatAt",
          "notificationChannels"
        ]
      },
      "UpdateUserSettingsDocDto": {
        "type": "object",
        "properties": {
          "heartbeatIntervalDays": {
            "type": "number",
            "example": 60,
            "minimum": 7,
            "maximum": 365
          },
          "isHeartbeatActive": {
            "type": "boolean",
            "example": true
          },
          "notificationChannels": {
            "type": "array",
            "example": [
              "email"
            ],
            "items": {
              "type": "string",
              "enum": [
                "email"
              ]
            }
          }
        }
      },
      "HeartbeatTriggerDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "manual",
            "enum": [
              "manual",
              "auto"
            ]
          }
        }
      },
      "GuardianDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cku6l6xhc0000q5l5zj9m1xyz"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "relationship": {
            "type": "string",
            "example": "Friend"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "relationship"
        ]
      },
      "HeartbeatGuardianResponseDto": {
        "type": "object",
        "properties": {
          "priority": {
            "type": "number",
            "example": 1
          },
          "guardian": {
            "$ref": "#/components/schemas/GuardianDetailsDto"
          }
        },
        "required": [
          "priority",
          "guardian"
        ]
      },
      "AddHeartbeatGuardianDto": {
        "type": "object",
        "properties": {
          "guardianId": {
            "type": "string",
            "example": "cku6l6xhc0000q5l5zj9m1xyz",
            "description": "ID of the guardian to assign"
          },
          "priority": {
            "type": "number",
            "example": 1,
            "description": "Order in which the guardian should be contacted (1 is highest priority)"
          }
        },
        "required": [
          "guardianId",
          "priority"
        ]
      }
    }
  }
}